Progress on intergrating Newton Dynamics
Minimal integration for Newton Dynamics. The action happens in
SatActiveArea and the newly added FtrNewtonBody.

So far, SatActiveArea holds a NewtonWorld, and FtrNewtonBody
holds a pointer to a NewtonBody. FtrNewtonBody is a Magnum
Feature that can be added to Object3Ds.

SatActiveArea now has a update_physics() function, called by the
OSPMagnum Magnum Application with a constant 1/60 timestep. It
updates the Newton world, and sets transformations of objects
that have a FtrNewtonBody.

Magnum matrices seem to be compatible with Newton ones, in
column-major order.

A FtrNewtonBody is added to each instantiated part for now, and
WASD camera control was added to SatActiveArea.

FIXME: Mesh resources have to be destructed in OpenGL thread