Add and Test new Icosahedron generator
IcoSphereTree was copied over from urho-osp, but it was so
terrible that it just had to be rewritten. In fact, the old one
didn't even generate a proper icosahedron, it was slightly
deformed. The new IcoSphereTree generates a proper icosahedron
and prints vertices to cout, where they can be verified.

Notable changes

* Added definitions for previously empty classes
* Added SatActiveArea load_func_add for SatPlanet
* Fixed Satellite m_loadedBy being uninitialized
* A few other changes