Progress on rewriting Planet Renderer from urho-osp
PlanetGeometryA is based on PlanetWrenderer from urho-osp. The
name would probably change to something like PlanetMeshGenA very
soon.

It copies over many of the same features, such as having very
little reallocations, sharing vertices, and all the planned-but-
not-implemented features like proper normal calculations. The
only real change is renamed variables, and separating OpenGL
functions to another class.

It's raviolied in this arrangement:

* IcoSphereTree manages a subdividable icosahedron
* PlanetGeometryA creates patches of triangles (chunks) over the
  IcoSphereTree's triangles for rendering and collision
* FtrPlanet interfaces PlanetGeometryA's buffers with Magnum/GL
  for rendering, and NewtonDynamics for collision

So far, PlanetGeometryA creates 10 chunks and is drawn by
FtrPlanet. That is all that's implemented so far.