Add ENTT submodule, Change to C++17, Remove Magnum SceneGraph
Some very drastic changes it seems. Magnum SceneGraph is not
ideal for games. It's good for a tech demo featuring an advance
new rendering technique, but not a for game engine for space
flight with all the described machines and nonsense. It has been
kind of inconvenient and old-fashioned to access 'features' of
objects because of the way it's structured.

Some sort of ECS scene graph seems to be the way to go here, and
ECS is the preferred way to structure games in this era. Many of
the planned features of OSP are better implemented in ECS, such
as Wiring and Trajectories, more on this somewhere else.

Everything that uses the SceneGraph has been commented out or
removed, and Active Area's test box render has been uncommented.