Rename a bunch of classes/structs and Progress on Machines
Names like "PartPrototype" violates the naming scheme used for
most of the project. I'm not quite sure what it's formally
called, but "adjective-last, base class first, or important info
first" is preferred. This should make it easier to read
(and also sort) sequences of variables.

ie: this is kind of hard to read:
* RedCar
* YellowCar
* GreenCar

compared to this:
* CarRed
* CarYellow
* CarGreen

It's almost like a namespace or category, or just use a
namespace in some cases.

A "MachinePrototype" and "MachineBlueprint" was going to be
added, but then that implies that those two will be different
kinds of Machine classes. Along with a few other additions,
everything related to Prototypes and Blueprints were renamed.

Notable changes

* Updated SysMachine: added update_physics and update_sensor
  virtual functions
* Added Machine Rocket: doesn't do anything yet
* Added some information on update loop in README
* Added maps and vectors of AbstractSysMachine in ActiveScene to
  manage and call updates
* Renamed classes and structs:
  * PartPrototype     -> PrototypePart
  * ObjectPrototype   -> PrototypeObject
  * VehicleBlueprint  -> BlueprintVehicle
  * PartBlueprint     -> BlueprintPart
  * WireBlueprint     -> BlueprintWire
* Added some test code in SturdyImporter to read glTF extra data
  for Machine configs
* A few other changes