Add SysDebugObject and FunctionOrder
ECS makes it slightly harder to make quick changes or add extra
functionality to individual entities: an entirely new system has
to be created. This isn't beneficial when there's only 1 or 2
entities using the new component. SysDebugObjects works around
this by introducing a component that holds a unique_ptr to a
polymorphic object.

There's "debug" in its name to discourage its use for anything
serious.

The way updates functions are being called in ActiveScene is
starting to get really messy, so the FunctionOrder class was
implemented. It's not very complete yet, but it's a class where
std::functions can be added to it, and it will automatically
order them correctly based on identifiers and before/after
rules. It handles cases like "this function must be called after
the wire update, but before the physics update."

I'm not that sure of how good of an idea this is, but it should
be very extension-friendly. The idea came from frustrations with
other game engines.

Notable changes:

* Added list_upd command in CLI
* Added SysDebugObject, read above
* Added FunctionOrder, read above
* Added delta time parameter to update_physics of Machines
* A few other changes