Implement FunctionOrder and Modify existing classes to use it
Most of the update-related function calls in ActiveScene was
removed, and instead replaced by m_updateOrder: an UpdateOrder
(aka: FunctionOrder<void(void)>). Systems classes now use a
FunctionOrderHandle to specify update functions, automatically
order them, and also manage lifetime.

FunctionOrder uses a list of reference-counted std::functions
with before/after rules. As mentioned in the previous commit,
it handles cases like "this function must be called after the
wire update, but before the physics update." Each registered
function call has strings: m_name, m_after, and m_before.

* m_name: for identification
* m_after: Specify that this function call must be placed after
  the function with this name
* m_before: Specify that this function call must be placed
  before the function with this name

FunctionOrderHandles control the lifetime of a function call in
the FunctionOrder. It is used as a class member so it would
automatically destruct with the class. The function calls are
reference counted, so it's like a shared_ptr. Removing the
call from the FunctionOrder isn't implemented yet.

FunctionOrderHandle members are initialized like this:

    m_updatePhysics(scene.get_update_order(),
            "mach_rocket", "wire", "physics",
            std::bind(&SysMachineRocket::update_physics, this))

The list_upd CLI command should output this:

Update order:
* mach_usercontrol
* wire
* mach_rocket
* physics

Notable changes:

* Modified every Sys* class to use FunctionOrder, read above
* Added more comments everywhere
* Set a bunch of functions to constexpr
* A few other changes