Restructure pretty much everything
This is simply moving code to where it makes the most sense.
It's probably best if all the code for a specific feature are
close together in the file system, instead of scattered across
different directories with a bunch of other stuff.

Rocket and UserControl were put into the new adera directory,
as they're game-focused and not really core features. Again, the
name "adera" was just the name of the stop the bus was on while
this project was created. It sounds better than "default" and
kind of describes its contents as a separate optional component;
anyways, this is an open source project.

Planet stuff was moved into src/planet-a. Yeah, planets are not
core features. There are many other ways to implement large
astronomical objects, maybe voxels, so it's probably best not to
hard-code the originally-known-as-PlanWren "Planet Renderer A"
into this thing.

maybe consider more namespaces.

For some reason, Packages are stored in the Universe class. So
instead, the OSPApplication class was created store the
Universe, Packages, and maybe any other part of OSP that doesn't
need an OpenGL context.

Those activate_something functions in SatActiveArea were moved
into their own class: SysVehicle. SysVehicle will deal with
vehicles in an ActiveArea.

Notable changes:

* Removed activate functions from SatActiveArea
* Added SysVehicle
* Added OSPApplication
* Moved src/osp/Active/Planet to src/planet-a
* Moved src/osp/Active/Machines to src/adera/Machines
* Moved src/types.h to src/osp/types
* Updated include paths
* A few other changes