Clean up SysMachine ECS implementation
For some reason, SysMachine managed the vector of machines, and
functioned like another ECS tacked onto the side of ENTT. The
main reason for this, is to allow multiple of the same machine
to be added to a part, which ENTT isn't made for. Machines are
tracked by a part through a CompMachine at its root entity.
CompMachine uses a linked list to track machines, and this is
inconvinent when accessing Machines via an index.

This commit changes this, by making machines into ENTT
components. This means that only one machine of a kind can be
added to an entity. Entities are very cheap, and parts are made
up of multiple entities already, so why not just use multiple
entities for multiple machines?

To keep track of what Machines a part has, CompMachine now uses
a vector of PartMachine structs: an entity and a map iterator to
a SysMachine.

Other Notable changes:

* Added empty SysPlanetA files
* Added WireMachineConnection struct, intended to better keep
  track of wire connections in the future
* Renamed AbstractSysMachine to ISysMachine
* Renamed WireElement to IWireElement
* A few other changes