Improve Planet vertex sharing
This commit adds m_vrtxSharedIcoCorners to PlanetGeometryA. This
keeps track of which shared vertices are associated with
individual vertices of the IcoSphereTree, making it really easy
to access and share.

If two chunks are generated that don't neighbour each other, but
their corners are touching, then they are now able to share that
corner vertex.