Implement Vehicle Deactivation
This new 'deactivation' function updates the Satellite positions of
activated Vehicles to their position in the Active Area/Scene.

Deactivation is called when the Active Area is disconnected from the
SysAreaAssociate, which is done once the Magnum Application window is
closed.

Start the application, move the vehicles, close the window, then start
the application again. The vehicle positions should stay persistent.

Saving rotations and velocity is not yet implemented.

Other Notable changes

* Added separate entity definition for Satellite, instead of a 'using'
  alias to 'entt::entity'. This prevents unintentional conversion to
  ActiveEnt, which is entt::entity
* Removed osp::universe::traj namespace. Trajectories are instead
  prefixed with Traj
* A few other changes