Implement Gravity "Force Field" for Planets
An ACompFFGravity can be added to an entity to make it attract rigid
bodies using universal gravitation. This force field behaviour is
inspired from Blender, and more stuff like filters or layers will be
added soon.

SysPlanetA was modified to add gravity to activated planets. It's now
possible to launch the spamcan vehicles into orbit, as well as
rendezvous them.