Add slightly more general AssetImporter
`SturdyImporter` works great for parts, but when creating general 3D
effects, it is limiting since it is only capable of importing images
that are directly referenced by the sturdy file. `AssetImporter` is
mostly a copy of `SturdyImporter`, but with general functions to load
and compile meshes and image textures. These functions can be used to
load in any image texture for use in shaders or other non-part uses.
Meshes are still loaded via `load_sturdy()`, but it no longer ignores
meshes which don't belong to `part_*` nodes. Future node types will
be explictly handled by the loading loop in `load_sturdy()`, but for
now `part_` is the only one necessary.

Right now, it's possible to load the same resource multiple times,
since the texture names from a glTF file are different from the
actual filepath to that mesh. Avoid doing this for now, in the future
we'll find some way to make sure each asset can only have one
identifier.