Create SysExhaustPlume
SysExhaustPlume is a system that handles the exhaust effects of
MachineRocket parts. A MachineRocket, upon instantiation, searches the
scene hierarchy underneath itself for nodes whose name begins with
`fx_plume_` and gives them a `ACompExhaustPlume`. This component stores
the ID of the owning entity, and uses its throttle input to determines
whether or not to print an annoying console message for informational
purposes.