Add plume graphics
Temporary system to attach visual effect to plume. For now it just
attaches a mesh with a placeholder diffuse texture, but next commit will
introduce transparency and shader effects.

Detailed list of changes:

- Added `CompVisibleDebug` component, which is used by SysDebugRender to
enable or disable the rendering of a certain entity

- Added `PlumeEffectData` resource which is effectively a
"PrototypePlume" that stores the information necessary to instantiate a
plume effect. Has some suspiciously useful looking members commented
out, wonder what they're for?

- Added `load_plume()` to AssetImporter; functions just like
`load_part()` but for loading `PlumeEffectData` resources

- SysMachineRocket, upon instantiating a new MachineRocket, searches the
node's hierarchy for nodes whose name begins with `fx_plume_` and uses
the name to try to instance the `ACompExhaustPlume` and
`CompDrawableDebug` components