Use ActiveScene's SysDebugRender, SysPhysics, and SysWire as Dynamic Systems
A few systems were members of ActiveScene for development reasons, as it
was planned to have a static systems, where ActiveScene was a tuple-like
template class (ActiveScene<SysA, SysB, ...>). This was dropped as there
wouldn't be any performance benefit, and would make the code confusing.

This commit will decouple SysDebugRender, SysPhysics (aka. SysNewton),
and SysWire from ActiveScene, and also clean up the includes in
ActiveScene.h. The systems are instead registered in main.cpp.