Add transparent rendering, exhaust effects
To add some fancy exhaust effects, we need to be able to draw
transparent objects. The rendering behavior has been separated out into
three functions: a function `draw_element()` that draws an arbitrary
`CompDrawableDebug`, a function `draw_group()` that calls that function
on every element of an EnTT observer type, and the original `draw()`
function which is called to draw the whole scene. This makes it easy to
draw transparent objects, since the the opaque and transparent objects
can be separated based on whether they possess a `CompTransparentDebug`
component. The opaque objects are drawn first, and then the transparent
backfaces, and finally the transparent front faces.

This also adds a fancy exhaust plume shader effect, PlumeShader. A
`plume_` object encodes information about its visual effect in its node
extras, which are read into a `PlumeEffectData` resource, as before.
Now, however, this object stores information that is used to set a
`PlumeShader` instance's uniform variables. This shader effect is stored
in the drawable component like usual and used when rendering the plume.