Add ShipResources
Adds a number of tools for storing and handling consumable on-ship
resources such as fuel, battery charge, etc. ShipResourceType provides a
definition for a specific type of resource and how it's stored
numerically (resource quantities will be represented by 64-bit unsigned
integers; see comments in source for detail). MachineContainer will
represent a container to store such resources. Not implemented anywhere
in the program, yet.