Rework Floating Origin and Rigid Body positioning
All Floating-origin code was removed from ActiveScene and SysNewton.
Translation is performed in SysAreaAssociate, and is called from
DebugCameraController.

The ActiveArea satellite is not yet moved. So when exiting the window,
the positions of the vehicles will be saved near the origin (they will
be incorrect).

ACompTransform no longer has m_enableFloatingOrigin. Floating origin
can now be enabled by adding an ACompFloatingOrigin.

This commit also adds a generalized 'controlled entity' option to
ACompTransform, for when a transform is updated each frame by a certain
system. For example, Entities with ACompRigidBody are controlled by
SysPhysics. Other possible uses will be for Distant objects, where an
entity's position will be constrained to a Satellite in the universe.