Add per-object machines, minimal fuel consumption
The ability for sub-part objects to own their own machines has a number
of benefits and will facilitate future features like integral RCS
thrusters, or realistic resource mass distribution. Upon import, parts
are now recursively scanned for machines and PrototypeMachines are
stored in PrototypeObjects instead of PrototypePart. BlueprintObjects
also store BlueprintMachines, which are used for unique configuration
options, e.g. specifying the type or quantity of fuel in a tank,
choosing whether a piece of equipment is retracted or extended
initially, etc.

This commit also sets up a bare minimum fuel consumption situation where
all container machines are automatically filled with a generic "fuel",
and all rockets are automatically connected to the large fuel tank (the
vessel is currently searched for the first non-empty fuel tank).
Eventually some sort of blueprint system will allow for the programmatic
connecting of fuel tank to engines. The engine configuration parameters
(currently just max thrust and specific impulse) are read in and used to
compute the fuel mass burn rate.