Add realtime CoM update, fuel tank physics
With engines now drawing resources from tanks, this commit has
`MachineContainer`s updating their `ACompMass` component for use by the
physics system to compute the ship mass distribution.

The old `ACompCollisionShape` has been split into two new components:
`ACompShape`, and `ACompCollider`. This allows the concept of shapes to
be applied beyond collisions; in this case it is used to define the
volume associated with a fuel/cargo container.

This new arrangement allows for inertia to be calculated much more
homogeneously. Since inertia requires not only masses but mass
distributions, the dry mass of parts (in the form of ACompMass) is
distributed to the part's collision volume on instantiation. These
collider shapes are used together with the mass to compute the moment of
inertia of the part overall, assuming the mass is distributed uniformly
throughout the volumes (for now). This arrangement also naturally
handles the same calculation for MachineContainers, since containers
(with their shape and mass components) look identical to massive
collision volumes as far as the physics system is concerned.

A new member, `bool DataRigidBody::m_inertiaDirty` may be set by
external systems to signal to the physics system that the ship's mass
has changed or shifted, requiring the moments of inertia to be
recalculated. This allows for the realistic shifting of the ship's
center of mass as resources (e.g. fuel) are consumed.

Finally, in order for rocket exhaust plumes to properly shut off when
engines run out of fuel, `MachineRocket`s now possess a
`current_output_power()` function which returns the output power of the
engine between [0,1]. SysExhaustPlume uses this value instead of
throttle input to determine how the plume should appear.