Finally Fix headache-inducing bug with planet corner vertex sharing
The issue was about how 'shared corners' are deleted in PlanetGeometryA.

Vector m_vrtxSharedIcoCorners keeps track of which shared vertices are
associated with vertices on the IcoSphereTree, aka: each of the 3
corners of each chunk.

When a chunk is removed, it checks if its 3 corners have no more users;
and if so, remove them from m_vrtxSharedIcoCorners.

The bug happens when a large chunk is neighbouring a small chunk, where
the small chunk's corners are shared with the large chunk's sides. If
the small triangle is deleted, then it will not delete the corners that
were shared with the larger chunk. When the large chunk is deleted, the
only corners it will remove from m_vrtxSharedIcoCorners is 3 of its own,
and the vertices that were formerly the small triangle's corners are
left undeleted.

The fix was to change how shared corners are deleted, and make the
chunk_remove function check if every deleted shared vertex is a corner,
instead of just the 3 from the current chunk.

m_vrtxSharedIcoCornersReverse was added to map shared vertices back to
an index in m_vrtxSharedIcoCorners.

Other changes:

* Added debug_verify_state functons to PlanetGeometryA and
  IcoSphereTree. These functions check stuff and make sure the state of
  the entire class is valid, used to narrow down on the bug.
* Moved 'raise vertices by shared count' into its own function
* Clean up here and there and hopefully improve readability