Improve GL Context-dependent resource management
We've been plagued by crashes on exit, which occur due to the
order-sensitive nature of freeing resources. This fixes the central
issue by tracking resources associated with the OpenGL context inside
OSPMagnum (which owns the GL context) rather than in the normal resource
packages. The resources are then properly destructed when the OSPMagnum
instance is destroyed, without affecting the OSPApplication data.