Implement Physical RCS
Add ACompMass to store part masses, which is read by AssetImporter from
sturdy parts. Part masses are stored in PrototypeParts, and used to
instantiate the ACompMass upon vehicle istantiation

Added `find_transform_rel_rigidbody_ancestor()`, which is similar to
`find_rigidbody_ancestor()` but returns the transformation that takes
the ancestor to the child rather than the ancestor itself.

Added ACompRigidbodyAncestor; it's filled automatically by
`try_get_or_find_rigidbody_ancestor()` and stores both the ancestor
entity itself and the relative transform between the two.
SysMachineRocket and SysRCSController access it via
`try_get_or_find_rigidbody_ancestor()`.

Add `try_get_or_find_rigidbody_ancestor()` function which `try_get()`s the
rigidbody component from an entity if the entity is valid, or searches
for and assigns the correct entity it it hasn't yet been found

Add `get_rigidbody_CoM()` which fetches the center of mass of a
rigidbody from Newton

Add `compute_body_CoM()` which computes and returns the center of mass
of a hierarchy subtree from the mass and transformation of all nodes
within the tree

Gave ACompNwtBody a center of mass member so that external code can
access a RB's center of mass without needing to request it from Newton.
This is currenly initialized by a call to `compute_rigidbody_CoM()` in
`SysNewton::create_body()`

Implemented physically accurate MachineRocket thrust. Thrust direction
is calculated based on the location of the rocket engine on the
rigidbody and the rigidbody orientation, and the world-space thrust and
torque are applied to the rigidbody

Updated SysMachineRCSController to read in RB CoM, no longer
hacked/hard-coded in

Updated DebugObject to center camera on the active vessel's center of
mass rather than its root node

Fixed issue where switching between vessels that didn't have
MachineUserControl caused a crash

Vehicle modification code invalidates all ACompRigidbodyAncestors upon
splitting a vehicle into multiple islands

Added a collider to the RCS block mesh