Add new rendering system, offscreen rendering
The previous SysDebugRender has been removed, and replaced with a
more idiomatic SysRender. Entities to be drawn in the renderer are given
an `ACompVisible` component, which indicates that the entity should be
processed by graphics code. Other components like `ACompOpaque` and
`ACompDiffuseTex`, `ACompMesh`, etc. are used to attach rendering
properties to the entity. An `ACompShader` stores a function pointer to
the shader draw function, which consumes whichever components are needed
to fill its uniform values before drawing the mesh.

Multiple rendering schemes are supported by creating `RenderPipeline`
resources which store RenderOrder_t objects which can be filled with
render passes. An agent with component `ACompRenderingAgent` is used to
fetch an entity storing an `ACompRenderTarget`, and a camera is
retrieved from the agent's `ACompPerspective3DView` component. The
ActiveScene finally retrieves the renderer by looking up the ID
contained in the agent's `ACompRenderer` component, which is then used
to draw the scene. Render passes consume components like `ACompOpaque`
and `ACompTransparent` to collect entities into the correct stage of
rendering.

A default render target is created automatically by SysRender, and is
rendered to the screen via a fullscreen triangle when the frame is
finally displayed.